refactor user UI loop
