+++ src/datautils/CSVHandler.java
@@ -1,0 +1,5 @@
+package datautils;
+
+public abstract class CSVHandler extends DataHandler {
+    protected abstract void updateCSV();
+}
