+++ src/datautils/DataHandler.java
@@ -1,0 +1,8 @@
+package datautils;
+
+/**
+ * Base class for all classes that read from / write to disk.
+ */
+public abstract class DataHandler {
+    public static final String DATADIR = "../data";
+}
